move graphite/ inside client/
