+++ client/graphite/config/config_test.go
