+++ client/graphite/config/testdata/graphite.good.yml
